rename: Move next.config.js helpers to new folder
The legacyFilterSlugs.json file is left as is because PHP expects it to live there. After we update the submodule in PHP, we can use the new location of the json file and drop this file.
