+++ frontend/next-config/redirects/legacy-filter-slugs.json
@@ -1,0 +1,197 @@
+{
+   "tagToItemType": {
+      "AC_Inlets": "Ports",
+      "Adapters": "Power_Adapters",
+      "Adhesive_Strips": "Adhesives",
+      "Bluetooth_Boards": "Wireless_Boards",
+      "Bumpers": "Case_Components",
+      "Card_Cages": "Case_Components",
+      "Charger_Boards": "Boards",
+      "Click_Wheels": "Buttons",
+      "Clips": "Brackets",
+      "Controllers": "Case_Components",
+      "DC-In_Boards": "Ports",
+      "Digitizers": "Screens",
+      "Displays": "Screens",
+      "Dock_Connectors": "Cables",
+      "Docks": "Ports",
+      "Earbuds": "Accessories",
+      "Flash": "Case_Components",
+      "Front_Panels": "Screens",
+      "Hard_Drive_Brackets": "Brackets",
+      "Hard_Drive_Enclosures": "Case_Components",
+      "Hard_Drive_Kits": "Kits",
+      "Hard_Drives_\\(PATA\\)": "Hard_Drives",
+      "Hard_Drives_\\(SATA\\)": "Hard_Drives",
+      "Hinges": "Case_Components",
+      "I/O_Board": "Boards",
+      "I": "Boards",
+      "Induction_Coil": "Charging_Coils",
+      "Inverters": "Boards",
+      "LCDs": "Screens",
+      "Lenses": "Cameras",
+      "Lightning_Connector": "Cables",
+      "Logic_Boards": "Motherboards",
+      "Magnets": "Case_Components",
+      "Memory_Maxxer_Kits": "RAM",
+      "Midframe": "Case_Components",
+      "Power_Jacks": "Ports",
+      "Rubber_Feet": "Case_Components",
+      "SD_Card": "Hard_Drives",
+      "SSD_Upgrade_Kits": "Hard_Drives",
+      "SSDs": "Hard_Drives",
+      "Straps": "Case_Components",
+      "Styluses": "Accessories",
+      "Test_Cables": "Cables",
+      "Timing_Control_Boards": "Boards",
+      "USB_Boards": "Boards",
+      "Wi-Fi_Boards": "Wireless_Boards",
+      "Wireless": "Wireless_Boards"
+   },
+   "tagsWithNoItemType": [
+      "Audio",
+      "Bearings",
+      "Belts",
+      "Computers",
+      "Consumables",
+      "DC-to-DC_Boards",
+      "Filters",
+      "Gaskets",
+      "Gripping",
+      "iFixit_Exclusives",
+      "Inspection",
+      "Lasers",
+      "Manuals",
+      "Microsoft_ASP",
+      "Modems",
+      "Motors",
+      "PC_Cards",
+      "PCIe",
+      "Propellers",
+      "Prying_and_Opening",
+      "Pulleys",
+      "Pumps",
+      "Retail_Display",
+      "RJ-11_Boards",
+      "SD_Card_Slots",
+      "Software",
+      "Teardown_Case",
+      "Textile_Repair",
+      "Thermal_Pads",
+      "Vaude"
+   ],
+   "toolLegacyRoutes": [
+      "AC_Inlets",
+      "Accessories",
+      "Adapters",
+      "Adhesive_Pads",
+      "Adhesive_Strips",
+      "Antennas",
+      "Audio",
+      "Batteries",
+      "Bearings",
+      "Belts",
+      "Bluetooth_Boards",
+      "Bumpers",
+      "Buttons",
+      "Cables",
+      "Cameras",
+      "Card_Cages",
+      "Case_Components",
+      "Charger_Boards",
+      "Clearance",
+      "Click_Wheels",
+      "Clips",
+      "Computers",
+      "Controllers",
+      "DC-In_Boards",
+      "DC-to-DC_Boards",
+      "Daughterboards",
+      "Digitizers",
+      "Display_Components",
+      "Displays",
+      "Dock_Connectors",
+      "Docks",
+      "Earbuds",
+      "Fans",
+      "Filters",
+      "Flash",
+      "Front_Panels",
+      "Gaskets",
+      "Graphics_Cards",
+      "Hard_Drive_Brackets",
+      "Hard_Drive_Enclosures",
+      "Hard_Drive_Kits",
+      "Hard_Drives",
+      "Hard_Drives_(PATA)",
+      "Hard_Drives_(SATA)",
+      "Headphone_Jacks",
+      "Heat_Sinks",
+      "Hinges",
+      "I/O_Board",
+      "Ice_Maker",
+      "Induction_Coil",
+      "Inverters",
+      "Joysticks",
+      "Keyboards",
+      "Kits",
+      "LCDs",
+      "Lamp",
+      "Lasers",
+      "Lenses",
+      "Lightning_Connector",
+      "Logic_Boards",
+      "MagSafe_Boards",
+      "Magnets",
+      "Manuals",
+      "Memory_Maxxer_Kits",
+      "Microphones",
+      "Microsoft_ASP",
+      "Midframe",
+      "Modems",
+      "Motherboards",
+      "Motors",
+      "Optical_Drives",
+      "Organization",
+      "PCIe",
+      "PC_Cards",
+      "Ports",
+      "Power_Jacks",
+      "Power_Supplies",
+      "Processors",
+      "Propellers",
+      "Pulleys",
+      "Pumps",
+      "RAM",
+      "RJ-11_Boards",
+      "Retail_Display",
+      "Rubber_Feet",
+      "SD_Card",
+      "SD_Card_Slots",
+      "SIM",
+      "SSD_Enclosures",
+      "SSD_Upgrade_Kits",
+      "SSDs",
+      "Screen_Protectors",
+      "Screens",
+      "Screwdrivers_and_Nutdrivers",
+      "Screws",
+      "Sensors",
+      "Software",
+      "Speakers",
+      "Straps",
+      "Styluses",
+      "Switches",
+      "Teardown_Case",
+      "Test_Cables",
+      "Textile_Repair",
+      "Thermal_Pads",
+      "Timing_Control_Boards",
+      "Trackpads",
+      "USB_Boards",
+      "Vaude",
+      "Vibrators",
+      "Wi-Fi_Boards",
+      "Wireless"
+   ]
+}
